requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

anyio==4.6.2.post1
asyncio==3.4.3
azure-functions==1.17.0
binance-connector==3.9.0
bitfinex-api-py==3.0.3
certifi==2024.7.4
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.7
pycryptodome==3.21.0
pyee==9.0.4
python-decouple==3.8
python-telegram-bot==21.7
requests==2.31.0
sniffio==1.3.1
types-requests==2.31.0.20240406
types-urllib3==1.26.25.14
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
websockets==11.0.3

